ui: add static images to asset build step
During the `genassets` build + embed step, we were taking just the
output of the `db-console-ccl` or `db-console-oss` step which is just
a build.js file. This commit adds references to the image assets we
want bundled as well. This includes favicon.ico and everything in `./
assets` relative to the db-console build directory.

We disable content hashing in webpack in order to keep the filenames
static, which bazel requires. The impact should be minimal as we
rarely change these images so if they're cached forever, it's fine.

This change restores the favicon to the CRDB build and the nice image
that shows up in the background of the email signup bar.

The size of the final zipped bundle only differs by around 1MB and is
already 10MB in size.

Fixes: #117876
Epic: None

Release note (ui change): the favicon now renders properly for DB
Console along with other image files.